apps/ingest_server/requirements.txt
python-dotenv==1.1.1
fastapi==0.118.0
hera==5.25.1
uvicorn==0.37.0